Add checking for player's distance near portal
Implement IsNearPortal() to check the player's distance along portal forward direction, return true if the player has crossed the desired distance.

Note: This one is different from CheckForPortalCrossing() that it doesn't require the player camera to be within portal threshold, it just need to check the some distance away from the portal.